refactor:imageUploadSection 훅으로 로직 분리(ui-기능 책임분산)
